test: use 32-bit store for test finisher

+++ tests/simple_payload/simple_payload.c
@@ -7,12 +7,12 @@
 #define TEST_FINISHER_ADDR 0x100000
 #endif
 
-#define TEST_FINISHER ((volatile unsigned short *)TEST_FINISHER_ADDR)
+#define TEST_FINISHER ((volatile unsigned int *)TEST_FINISHER_ADDR)
 
 void _start(void)
 {
-    /* Write pass status; QEMU will exit immediately. */
-    *TEST_FINISHER = (unsigned short)0x5555;
+    /* Write pass status (low 16 bits = 0x5555). The upper 16 bits can be zero for exit code 0 */
+    *TEST_FINISHER = 0x5555;
 
     /* Should not reach here. */
     while (1) {
